EES-3751 Move permalink migration code to PermalinkMigrationService with Startup configuration to satisfy DI.
