+++ src/GovUk.Education.ExploreEducationStatistics.Publisher/Startup.cs
@@ -126,6 +126,15 @@ namespace GovUk.Education.ExploreEducationStatistics.Publisher
                         logger: provider.GetRequiredService<ILogger<QueueService>>()))
                 .AddScoped<IReleasePublishingStatusService, ReleasePublishingStatusService>()
                 .AddScoped<IValidationService, ValidationService>()
+                .AddScoped<IPermalinkMigrationService, PermalinkMigrationService>(provider =>
+                    new PermalinkMigrationService(
+                        contentDbContext: provider.GetRequiredService<ContentDbContext>(),
+                        blobServiceClient: new BlobServiceClient(
+                            GetConfigurationValue(provider, "PublicStorage")),
+                        storageQueueService: new StorageQueueService(
+                            GetConfigurationValue(provider, "PublisherStorage"),
+                            new StorageInstanceCreationUtil()
+                        )))
                 .AddScoped<IReleaseSubjectRepository, ReleaseSubjectRepository>(provider =>
                     new ReleaseSubjectRepository(
                         statisticsDbContext: provider.GetRequiredService<PublicStatisticsDbContext>(),
